requirements.txt
openai
langchain
pinecone-client
unstrucutred
tiktoken